+++ lib/model/item_model.dart
@@ -1,10 +1,14 @@
 class Item {
   final String name;
   final String unit;
-  final int price;
+  final double price;
   final String image;
 
-  Item({required this.name, required this.unit, required this.price, required this.image});
+  Item(
+      {required this.name,
+      required this.unit,
+      required this.price,
+      required this.image});
 
   Map toJson() {
     return {
